[iris] Move write_workdir_files and normalize_workdir_relative_path to runtime/env.py
These functions are runtime helpers, not bundle store concerns. Moved from
bundle.py to cluster/runtime/env.py and updated all import sites.
