+++ lib/iris/src/iris/cluster/runtime/process.py
@@ -34,8 +34,8 @@ from dataclasses import dataclass, field, replace
 from datetime import datetime, timezone
 from pathlib import Path
 
-from iris.cluster.bundle import BundleStore, write_workdir_files
-from iris.cluster.runtime.env import build_device_env_vars
+from iris.cluster.bundle import BundleStore
+from iris.cluster.runtime.env import build_device_env_vars, write_workdir_files
 from iris.cluster.runtime.profile import (
     build_memray_attach_cmd,
     build_memray_transform_cmd,
